Added templated nap::Attribute to ofParameter utilities for gui control

+++ core/src/nap/object.cpp
@@ -233,7 +233,7 @@ namespace nap
 	int Object::getChildIndex(Object& child)
 	{
 		auto children = getChildren();
-		for (int i = 0; i < mChildren.size(); i++) {
+		for (size_t i = 0; i < mChildren.size(); i++) {
 			if (children[i] == &child) return i;
 		}
 		return -1;
